IR-1820 WIP playmode changes
