+++ packages/engine/src/interaction/systems/InteractableSystem.ts
@@ -26,10 +26,8 @@ Ethereal Engine. All Rights Reserved.
 import { Not } from 'bitecs'
 import { MathUtils, Vector3 } from 'three'
 
-import { defineState } from '@etherealengine/hyperflux'
 import { WebLayer3D } from '@etherealengine/xrui'
 
-import { isClient } from '@etherealengine/common/src/utils/getEnvironment'
 import {
   getMutableComponent,
   hasComponent,
@@ -44,10 +42,8 @@ import { Engine } from '@etherealengine/ecs/src/Engine'
 import { Entity } from '@etherealengine/ecs/src/Entity'
 import { defineQuery } from '@etherealengine/ecs/src/QueryFunctions'
 import { defineSystem } from '@etherealengine/ecs/src/SystemFunctions'
-import { createUI } from '@etherealengine/engine/src/interaction/functions/createUI'
 import { getState } from '@etherealengine/hyperflux'
 import { CallbackComponent } from '@etherealengine/spatial/src/common/CallbackComponent'
-import { createTransitionState } from '@etherealengine/spatial/src/common/functions/createTransitionState'
 import { EngineState } from '@etherealengine/spatial/src/EngineState'
 import { InputComponent } from '@etherealengine/spatial/src/input/components/InputComponent'
 import { InputPointerComponent } from '@etherealengine/spatial/src/input/components/InputPointerComponent'
@@ -58,33 +54,21 @@ import { RigidBodyComponent } from '@etherealengine/spatial/src/physics/componen
 import { HighlightComponent } from '@etherealengine/spatial/src/renderer/components/HighlightComponent'
 import { VisibleComponent } from '@etherealengine/spatial/src/renderer/components/VisibleComponent'
 import { BoundingBoxComponent } from '@etherealengine/spatial/src/transform/components/BoundingBoxComponents'
-import {
-  DistanceFromCameraComponent,
-  DistanceFromLocalClientComponent
-} from '@etherealengine/spatial/src/transform/components/DistanceComponents'
-import { EntityTreeComponent } from '@etherealengine/spatial/src/transform/components/EntityTree'
+import { DistanceFromCameraComponent } from '@etherealengine/spatial/src/transform/components/DistanceComponents'
 import { TransformComponent } from '@etherealengine/spatial/src/transform/components/TransformComponent'
 import { TransformSystem } from '@etherealengine/spatial/src/transform/systems/TransformSystem'
 import { XRUIComponent } from '@etherealengine/spatial/src/xrui/components/XRUIComponent'
 import { AvatarComponent } from '../../avatar/components/AvatarComponent'
 import { InteractableComponent, XRUIActivationType, XRUIVisibilityOverride } from '../components/InteractableComponent'
-import { gatherAvailableInteractables, inFrustum } from '../functions/interactableFunctions'
-
-export const InteractableState = defineState({
-  name: 'InteractableState',
-  initial: () => {
-    return {
-      /**
-       * all interactables within threshold range, in view of the camera, sorted by distance
-       */
-      available: [] as Entity[]
-    }
-  }
-})
+import {
+  gatherAvailableInteractables,
+  inFrustum,
+  InteractableState,
+  InteractableTransitions
+} from '../functions/interactableFunctions'
 
 //TODO get rid of the query.exit and put it in component as unmount useEffect return
 //TODO get rid of this map eventually and store on the component instead
-export const InteractableTransitions = new Map<Entity, ReturnType<typeof createTransitionState>>()
 
 const xrDistVec3 = new Vector3()
 
@@ -98,9 +82,10 @@ const updateXrDistVec3 = (selfAvatarEntity: Entity) => {
 
 export const onInteractableUpdate = (entity: Entity) => {
   const selfAvatarEntity = AvatarComponent.getSelfAvatarEntity()
-  if (!selfAvatarEntity) return
-
   const interactable = getComponent(entity, InteractableComponent)
+
+  if (!selfAvatarEntity || !interactable || interactable.uiEntity == UndefinedEntity) return
+
   const xrui = getComponent(interactable.uiEntity, XRUIComponent)
   const xruiTransform = getComponent(interactable.uiEntity, TransformComponent)
   const boundingBox = getOptionalComponent(entity, BoundingBoxComponent)
@@ -176,24 +161,6 @@ export const onInteractableUpdate = (entity: Entity) => {
   })
 }
 
-/**
- * Adds an interactable UI to the entity if it has label text
- * @param entity
- */
-const addInteractableUI = (entity: Entity) => {
-  if (!isClient || getState(EngineState).isEditor) return //no xrui in editor
-  const interactable = getComponent(entity, InteractableComponent)
-  if (!interactable.label || interactable.label === '') return //null or empty label = no ui
-
-  interactable.uiEntity = createUI(entity, interactable.label, interactable.uiInteractable).entity
-  setComponent(interactable.uiEntity, EntityTreeComponent, { parentEntity: Engine.instance.originEntity })
-
-  const transition = createTransitionState(0.25)
-  transition.setState('OUT')
-  InteractableTransitions.set(entity, transition)
-}
-
-const allInteractablesQuery = defineQuery([InteractableComponent])
 const interactableQuery = defineQuery([InteractableComponent, Not(AvatarComponent), DistanceFromCameraComponent])
 const hoverInputInteractablesQuery = defineQuery([InteractableComponent, InputComponent])
 
@@ -203,14 +170,6 @@ const execute = () => {
   gatherAvailableInteractablesTimer += getState(ECSState).deltaSeconds
   // update every 0.1 seconds
   if (gatherAvailableInteractablesTimer > 0.1) gatherAvailableInteractablesTimer = 0
-
-  // ensure distance component is set on all interactables
-  for (const entity of allInteractablesQuery.enter()) {
-    setComponent(entity, DistanceFromCameraComponent)
-    setComponent(entity, DistanceFromLocalClientComponent)
-
-    addInteractableUI(entity)
-  }
 
   for (const entity of interactableQuery.exit()) {
     if (InteractableTransitions.has(entity)) InteractableTransitions.delete(entity)
@@ -235,7 +194,7 @@ export const InteractableSystem = defineSystem({
 })
 
 const executeInput = () => {
-  if (getState(EngineState).isEditor) return
+  if (getState(EngineState).isEditing) return
 
   const inputPointerEntity = InputPointerComponent.getPointerForCanvas(Engine.instance.viewerEntity)
   if (!inputPointerEntity) return
